New tests for dedup, other changes too.
This was the commit that took forever because tox sets PYTHONHASHSEED
for you, and me not realizing that was the reason superdeduper deduce
kept giving non-deterministic output.  Eventually figured that and
set it to 0 in tox.ini.

Also, python2 vs python3: they have different random number 
generators, which results in different outputs for the different
versions, so had to add support for that in test framework.

Up next, doug-seq.